feat: add bcrypt

+++ src/modules/users/user.controller.ts
@@ -1,5 +1,6 @@
 import type { Request, Response, NextFunction } from "express";
 import * as userService from "./user.service.js";
+import { BadRequestError, NotFoundError } from "../../errors/errorTypes.js";
 
 export async function getUsers(
   req: Request,
@@ -13,11 +14,14 @@ export async function getUsers(
     next(error);
   }
 }
-
+// return incorrect user, example: userId is 3, input was 1, return 3 still
 export async function getUser(req: Request, res: Response, next: NextFunction) {
   try {
     const id = Number(req.params.id);
     const user = await userService.getUser(id);
+    if (!user) {
+      throw new NotFoundError("User does not exists in database");
+    }
     res.status(200).json(user);
   } catch (error) {
     next(error);
@@ -37,6 +41,24 @@ export async function createUser(
   }
 }
 
+export async function updateUser(
+  req: Request,
+  res: Response,
+  next: NextFunction,
+) {
+  try {
+    const id = Number(req.params.id);
+    const user = await userService.getUser(id);
+    if (!user) {
+      throw new NotFoundError("User does not exists in database");
+    }
+    const updatedUser = await userService.updateUser(id, req.body);
+    return res.status(200).json(updatedUser);
+  } catch (error) {
+    next(error);
+  }
+}
+
 export async function deleteUser(
   req: Request,
   res: Response,
@@ -44,6 +66,10 @@ export async function deleteUser(
 ) {
   try {
     const id = Number(req.params.id);
+    const user = await userService.getUser(id);
+    if (!user) {
+      throw new NotFoundError("User does not exists in database");
+    }
     await userService.deleteUser(id);
     res.status(200).json({ message: `User with id ${id} was deleted!` });
   } catch (error) {
